Ask for confirmation before performing sync
Syncing is an action with potential for large data loss.
By asking the user if he wants to proceed after
we have gathered information about which actions
we will perform, we reduce the risk of this.

Ideally, perhaps the interactivity should be handled
in bin/dptrp1, and also with an additional flag for
performing the action without asking for confirmation.
This would, however, require more refactoring. Since
asking for confirmation is so important, it was therefore
implemented in dptrp1.dptrp1 for now.